+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix the DICOM limit to say "Series" instead of "Studies" ([#1379](../../pull/1379))
 - Add token authentication option for DICOMweb ([#1349](../../pull/1349))
 - Use vips resize rather than reduce to increase speed ([#1384](../../pull/1384))
+- Added annotation to geojson endpoint and utility ([#1385](../../pull/1385))
 
 ### Bug Fixes
 - Fix an issue applying ICC profile adjustments to multiple image modes ([#1382](../../pull/1382))